requirements.txt
sentence-transformers<3.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
numpy>=1.25.2,<2.6.0
pandas>=1.5.3
torch>=2.0.0
streamlit
python-docx